Rename test module open_search_test_case to opensearch_test_case (#6779)
