+++ .isort.cfg
@@ -5,4 +5,4 @@ include_trailing_comma=1
 force_sort_within_sections=1
 case_sensitive=True
 combine_as_imports=1
-known_first_party=azul_test_case,open_search_test_case,indexer,health_check_test_case,app_test_case,sqs_test_case,service,docker_container_test_case,dynamodb_test_case
+known_first_party=azul_test_case,opensearch_test_case,indexer,health_check_test_case,app_test_case,sqs_test_case,service,docker_container_test_case,dynamodb_test_case
